Windows executable and updates
Added password functionality, handled a bunch of potential errors that can occur when networking, used a kill file to ensure all threads die once the client is closed. Handled a hangup that can occur on server side when a client abruptly closes the connection (I think. It's hard to replicate the conditions)